Moved helper functions used in plotting to scripts/plot_stats instead of hover_robot, reduced the max timesteps for A1GymEnv to 1000
